bugs

+++ plugins/simple_networking/lib/modules/wallet_api/data_sources/wallet_api_data_sources.dart
@@ -3139,7 +3139,7 @@ class WalletApiDataSources {
 
   Future<DC<ServerRejectException, InvestPositionResponseModel>> changePendingPrice({
     required String id,
-    required int price,
+    required double price,
   }) async {
     try {
       final response = await _apiClient.post(
